+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+ - Fix building for 32-bit Python on 64-bit Windows. #[77](https://github.com/PyO3/setuptools-rust/pull/77)
+
 ## 0.11.0 (2020-08-06)
 
  - Remove python 2 support. #[53](https://github.com/PyO3/setuptools-rust/pull/53)